// src/actions/action.types.ts
export type ActionKey =
  | 'OPEN_GLOBAL_SEARCH'

  | 'OPEN_DOCTOR_LIST'
  | 'OPEN_DOCTOR_DETAIL'
  | 'OPEN_BOOKING_SUCCESS'

  | 'OPEN_PHARMACY_LIST'
  | 'OPEN_PHARMACY'
  | 'OPEN_CART'

  | 'OPEN_LABS'
  | 'OPEN_LAB_TESTS'

  | 'OPEN_HOSPITALS'
  | 'OPEN_AMBULANCE'
  | 'OPEN_HOMECARE'

  | 'OPEN_RECORDS'
  | 'UPLOAD_RECORD'

  | 'OPEN_PROFILE'
  | 'OPEN_CALENDAR'
  | 'OPEN_HEALTH'
  | 'OPEN_NOTIFICATIONS'
  | 'ADD_MEDICINE_FROM_SEARCH'
  | 'INCREMENT_CART_ITEM'
  | 'DECREMENT_CART_ITEM'
  | 'OPEN_MY_MEDICINE_ORDERS'
  | 'OPEN_MY_LAB_TESTS'
  | 'OPEN_ADDRESS_SELECTOR'
  | 'OPEN_PROFILE_DETAILS'
  | 'OPEN_EDIT_PROFILE'
  | 'OPEN_DOCTOR_ONBOARDING'
  | 'OPEN_FAMILY_MEMBERS'
  | 'OPEN_ADD_FAMILY_MEMBER'
  | 'OPEN_ADDRESS_BOOK'
  | 'OPEN_CONSULTATIONS'
  | 'OPEN_MY_MEDICINE_ORDERS'
  | 'OPEN_REDEEM'
  | 'OPEN_DOCTOR_SETTINGS'
  | 'OPEN_MANAGE_AVAILABILITY'
  | 'OPEN_MANAGE_APPOINTMENTS'
  | 'OPEN_DOCTOR_PENDING_REQUESTS'
  | 'OPEN_DOCTOR_REVIEWS'
  | 'OPEN_DOCTOR_EARNINGS'
  | 'OPEN_PATIENT_CONSULTATION_DETAIL'
  | 'OPEN_DOCTOR_CALL'
  | 'OPEN_DOCTOR_CHAT'
  | 'OPEN_HELP'
  | 'OPEN_TERMS'
  | 'OPEN_ADD_ADDRESS'
  | 'OPEN_CONSULTATION_DETAIL'
  | 'OPEN_DOCTOR_FEEDBACK'
  | 'OPEN_MEDICINE_ORDER_DETAIL'
  | 'OPEN_LAB_TEST_DETAIL'
  | 'OPEN_LAB_TEST_DETAIL'
  | 'OPEN_CHECKOUT'
  | 'GO_BACK';